# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: Qoodle API ドキュメント
  version: 0.0.1
  title: Qoodle REST APIs
host: localhost:8080
basePath: /dbapi
...
schemes:
  - http
  - https
paths:
...
  /quiz:
    post:
      tags:
        - quiz
      summary: クイズ作成
      description: クイズ作成
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/QuizRequest'
#        - name: category
#          type: integer
#          in: formData
#          description: 0 一般クイズ（お絵描き解答含む）, 1 フリーテキストで出題, 2 16分割, 3 ぼかし
#        - name: point
#          type: integer
#          in: formData
#          description: 得点
#        - name: body
#          type: string
#          in: formData
#          description: 出題内容
#        - name: comment
#          type: string
#          in: formData
#          description: コメント（クイズ画面には非表示）
#        - name: img_url
#          type: string
#          in: formData
#          description: 画像を使うクイズの画像URL
#        - name: user_id
#          type: string
#          in: formData
#          description: ユーザーID
#        - name: user_name
#          type: string
#          in: formData
#          description: ユーザー名
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
...
  /quizset:
    post:
      tags:
        - quizset
      summary: クイズセット作成
      description: クイズセット作成
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/QuizsetRequest'
#        - name: quiz_ids
#          type: array
#          in: formData
#          description: クイズIDの配列
#        - name: user_id
#          type: string
#          in: formData
#          description: ユーザーID
#        - name: user_name
#          type: string
#          in: formData
#          description: ユーザー名
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
...
definitions:
  QuizRequest:
    type: object
    properties:
      category:
        type: integer
        description: 0 一般クイズ（お絵描き解答含む）, 1 フリーテキストで出題, 2 16分割, 3 ぼかし
      point:
        type: integer
        description: 得点
      body:
        type: string
        description: 出題内容
      comment:
        type: string
        description: コメント（クイズ画面には非表示）
      img_url:
        type: string
        description: 画像を使うクイズの画像URL
      user_id:
        type: string
        description: ユーザーID
      user_name:
        type: string
        description: ユーザー名
  QuizsetRequest:
    type: object
    properties:
      quiz_ids:
        type: array
        description: クイズIDの配列
      user_id:
        type: string
        description: ユーザーID
      user_name:
        type: string
        description: ユーザー名
...